+++ src/bcnavbit.rs
@@ -571,9 +571,9 @@ impl BCNavBit {
             // Encode key ephemeris parameters into D1/D2 message format
             eph1_data[0] = (svid as u32) << 24; // SVID 
             eph1_data[1] = (eph.toe as u32 - 14); // Adjust for BDT
-            eph1_data[2] = eph.sqrtA.to_bits();
-            eph1_data[3] = eph.ecc.to_bits();
-            eph1_data[4] = eph.i0.to_bits();
+            eph1_data[2] = (eph.sqrtA.to_bits() as u32);
+            eph1_data[3] = (eph.ecc.to_bits() as u32);
+            eph1_data[4] = (eph.i0.to_bits() as u32);
             
             // D2 message has different structure (10 words vs 9)
             eph2_data[0] = eph1_data[0]; // Copy SVID
@@ -581,7 +581,7 @@ impl BCNavBit {
             for i in 2..9 {
                 eph2_data[i] = eph1_data[i];
             }
-            eph2_data[9] = eph.omega_dot.to_bits(); // Additional field in D2
+            eph2_data[9] = (eph.omega_dot.to_bits() as u32); // Additional field in D2
             
             self.Ephemeris1[index] = eph1_data;
             self.Ephemeris2[index] = eph2_data;
